pyproject.toml
[project]
name = "code-scrap"
version = "0.2.0"
description = "Scrape vendor websites into Markdown knowledge bases for NotebookLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "pdfplumber>=0.10.0",
    "lxml>=4.9.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
codescrap = "codescrap.cli:main"
